pdf_summarizer/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
python-dotenv==1.0.0
openai==0.27.8
PyPDF2==3.0.1
gunicorn==20.1.0
flask-login==0.6.2
email-validator==2.0.0
